Utils: Move 'print_term' and 'Linear_polynomial' to 'utils'
