+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(uf)
 add_subdirectory(variable_order)
 add_subdirectory(terms)
 add_subdirectory(parser)
+add_subdirectory(utils)
 
 target_sources(yaga PRIVATE
     Conflict_analysis.cpp
