+++ src/utils/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_include_directories(yaga PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+
+target_sources(yaga PRIVATE
+        Utils.cpp
+        Linear_polynomial.cpp
+)
